Loop for created workflow in setup script
Currently when running the test script the setup script will break
the first time around. This is because it is expecting a newly created
workflow to definitely exist in the following `syncRecords` call. With
the new sync this might not be the case. Instead the setup script loops
a `syncRecords` call until it finds the correct workflow in the response.